+++ json/test/TSGeneratorTest.kt
@@ -26,7 +26,7 @@ class TSGeneratorTest {
       "interface Person {hello: SomeEnum; name: string}")
 
     expect(ts.render(SomeData::class)).toEqual( // language=TypeScript
-      "interface SomeData {age: number; birthDate?: string; bytes: Array<number>; field: keyof Person; id: MyId<SomeData>; " +
+      "interface SomeData {age: number; any: any; birthDate?: string; bytes: Array<number>; field: keyof Person; id: MyId<SomeData>; " +
         "list: Array<SomeData>; map: Record<string, Array<SomeData>>; name: string; other?: SomeData; status: SomeDataStatus; hello: SomeEnum}")
   }
 }
@@ -36,7 +36,7 @@ enum class SomeEnum { HELLO, WORLD }
 interface Person { val name: String; val hello get() = SomeEnum.HELLO }
 
 data class SomeData(override val name: String, val age: Int, val birthDate: String?, val id: MyId<SomeData>, val other: SomeData?,
-                    val list: List<SomeData>, val map: Map<LocalDate, Array<SomeData>>, val status: Status = Status.ACTIVE,
+                    val list: List<SomeData>, val map: Map<LocalDate, Array<SomeData>>, val any: Any, val status: Status = Status.ACTIVE,
                     val field: KProperty1<Person, *>, val bytes: ByteArray): Person {
   enum class Status { ACTIVE }
 }